+++ packages/node-fetch-server/global.d.ts
@@ -1,0 +1,3 @@
+interface ReadableStream<R = any> {
+  [Symbol.asyncIterator](): AsyncIterableIterator<R>;
+}