+++ packages/node-fetch-server/src/lib/read-stream.ts
@@ -1,9 +1,0 @@
-export async function* readStream(stream: ReadableStream<Uint8Array>): AsyncIterable<Uint8Array> {
-  let reader = stream.getReader();
-
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) break;
-    yield value;
-  }
-}